[CHORE] fix group invite
