+++ GumungGagye/GumungGagye/Assets.xcassets/SelectingBankImage/농협은행.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "Property 1=농협은행.png",
+      "filename" : "Property 1=농협은행.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Property 1=농협은행@2x.png",
+      "filename" : "Property 1=농협은행@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Property 1=농협은행@3x.png",
+      "filename" : "Property 1=농협은행@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
