+++ GumungGagye/GumungGagye/Assets.xcassets/SelectingBankImage/농협카드.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "Property 1=농협카드.png",
+      "filename" : "Property 1=농협카드.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "Property 1=농협카드@2x.png",
+      "filename" : "Property 1=농협카드@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "Property 1=농협카드@3x.png",
+      "filename" : "Property 1=농협카드@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
