+++ GumungGagye/GumungGagye/FirebaseQuery/FirebaseController.swift
@@ -329,6 +329,8 @@ class FirebaseController: ObservableObject {
                         return
                     }
                     self.fetchUsersData(userIDs: arrayData, completion: completion)
+                    
+                    
                 } else {
                     completion(nil)
                 }
@@ -352,9 +354,11 @@ class FirebaseController: ObservableObject {
             
             var users: [UserData] = []
             
+       
             for document in querySnapshot!.documents {
                 let data = document.data()
-                let user = UserData(data: data)
+                var user = UserData(data: data)
+        
                 users.append(user)
             }
